feat: Refactor ChatListScreen to use ChatListRoute (#132)
**Key changes:**
- `ChatListRoute.kt` created to manage `ChatListViewModel` and handle fetching the chat list.
- `ChatListScreen.kt`:
    - Now receives `chatRoomSummaries` and `onItemClick` as parameters.
    - The `ChatListItem` composable, previously in a separate file (`chat/components/ChatListItem.kt`), has been moved into `ChatListScreen.kt`.
- `ChatListItem.kt`: Deleted as its functionality is now within `ChatListScreen.kt`.
- `BottomTabNavigation.kt`: Updated to use `ChatListRoute` instead of `ChatListScreen`.
